Add files via upload
This dataset required significant preprocessing in order to get the correct set of features. This is the full code used to preprocess the dataset
Each feature is defined in terms of the number of seconds - so the units are uniform across the dataset, preventing the need for further normalization
Due to the large size of the dataset and limited computation power (this was done on my local machine), I performed the analysis only on 10,000 records. Given more computational power I would repeat this analysis with the full dataset.